Add frontend payment method for payolution, generate guaranteed invoice base class
