+++ Core/SessionHelper.php
@@ -70,15 +70,17 @@ class SessionHelper
     /**
      * Returns date of birth
      *
+     * @param string $sPaymentMethodName
+     *
      * @return string date of birth formatted for db (format 'Y-m-d')
      *
      * @since 1.2.0
      */
-    public static function getDbDateOfBirth()
+    public static function getDbDateOfBirth($sPaymentMethodName)
     {
         $oSession = Registry::getConfig()->getSession();
         $aDynvalues = $oSession->getVariable('dynvalue');
-        $sDateOfBirth = $aDynvalues['dateOfBirth'];
+        $sDateOfBirth = $aDynvalues['dateOfBirth' . $sPaymentMethodName];
         //To change the format string add a translation for another language on phrase app
         $oDateOfBirth = DateTime::createFromFormat(Helper::translate('wd_date_format_php_code'), $sDateOfBirth);
 
@@ -91,20 +93,21 @@ class SessionHelper
      * Sets date of birth
      *
      * @param string $sDbDateOfBirth formated for db (format 'Y-m-d')
+     * @param string $sPaymentMethodName
      *
      * @since 1.2.0
      */
-    public static function setDbDateOfBirth($sDbDateOfBirth)
+    public static function setDbDateOfBirth($sDbDateOfBirth, $sPaymentMethodName)
     {
         $oSession = Registry::getConfig()->getSession();
         $aDynvalues = $oSession->getVariable('dynvalue');
-        $aDynvalues['dateOfBirth'] = '';
+        $aDynvalues['dateOfBirth' . $sPaymentMethodName] = '';
 
         if ($sDbDateOfBirth !== self::DEFAULT_DATE_OF_BIRTH) {
             $oDateOfBirth = DateTime::createFromFormat(self::DB_DATE_FORMAT, $sDbDateOfBirth);
 
             if ($oDateOfBirth) {
-                $aDynvalues['dateOfBirth'] =
+                $aDynvalues['dateOfBirth' . $sPaymentMethodName] =
                     $oDateOfBirth->format(Helper::translate('wd_date_format_php_code'));
             }
         }
@@ -115,31 +118,34 @@ class SessionHelper
     /**
      * Returns true if a valid date of birth is available
      *
+     * @param string $sPaymentMethodName
+     *
      * @return bool
      *
      * @since 1.2.0
      */
-    public static function isDateOfBirthSet()
+    public static function isDateOfBirthSet($sPaymentMethodName)
     {
-        return self::getDbDateOfBirth() !== self::DEFAULT_DATE_OF_BIRTH;
+        return self::getDbDateOfBirth($sPaymentMethodName) !== self::DEFAULT_DATE_OF_BIRTH;
     }
 
     /**
      * Returns true if user is min $iAge years old, false if not or date of birth is not set
      *
-     * @param int $iAge
+     * @param int    $iAge
+     * @param string $sPaymentMethodName
      *
      * @return bool
      *
      * @since 1.2.0
      */
-    public static function isUserOlderThan($iAge)
+    public static function isUserOlderThan($iAge, $sPaymentMethodName)
     {
         $oSession = Registry::getConfig()->getSession();
         $aDynvalues = $oSession->getVariable('dynvalue');
 
         $oDateOfBirth =
-            DateTime::createFromFormat(Helper::translate('wd_date_format_php_code'), $aDynvalues['dateOfBirth']);
+            DateTime::createFromFormat(Helper::translate('wd_date_format_php_code'), $aDynvalues['dateOfBirth' . $sPaymentMethodName]);
 
         if (!$oDateOfBirth) {
             return false;
@@ -154,15 +160,17 @@ class SessionHelper
     /**
      * Returns phone
      *
+     * @param string $sPaymentMethodName
+     *
      * @return string
      *
      * @since 1.2.0
      */
-    public static function getPhone()
+    public static function getPhone($sPaymentMethodName)
     {
         $oSession = Registry::getConfig()->getSession();
         $aDynvalues = $oSession->getVariable('dynvalue');
-        $sPhone = $aDynvalues['phone'];
+        $sPhone = $aDynvalues['phone' . $sPaymentMethodName];
 
         return $sPhone ? $sPhone : '';
     }
@@ -171,14 +179,15 @@ class SessionHelper
      * Sets the phone number
      *
      * @param string $sPhone
+     * @param string $sPaymentMethodName
      *
      * @since 1.2.0
      */
-    public static function setPhone($sPhone)
+    public static function setPhone($sPhone, $sPaymentMethodName)
     {
         $oSession = Registry::getConfig()->getSession();
         $aDynvalues = $oSession->getVariable('dynvalue');
-        $aDynvalues['phone'] = $sPhone;
+        $aDynvalues['phone' . $sPaymentMethodName] = $sPhone;
 
         $oSession->setVariable('dynvalue', $aDynvalues);
     }
@@ -186,41 +195,46 @@ class SessionHelper
     /**
      * Returns true if a valid phone number is available or not needed
      *
+     * @param string $sPaymentMethodName
+     *
      * @return bool
      *
      * @since 1.2.0
      */
-    public static function isPhoneValid()
+    public static function isPhoneValid($sPaymentMethodName)
     {
-        return self::getPhone() !== '';
+        return self::getPhone($sPaymentMethodName) !== '';
     }
 
     /**
      * Returns saveCheckoutFields
      *
+     * @param string $sPaymentMethodName
+     *
      * @return string
      *
      * @since 1.2.0
      */
-    public static function getSaveCheckoutFields()
+    public static function getSaveCheckoutFields($sPaymentMethodName)
     {
         $oSession = Registry::getConfig()->getSession();
         $aDynvalues = $oSession->getVariable('dynvalue');
-        return $aDynvalues['saveCheckoutFields'];
+        return $aDynvalues['saveCheckoutFields' . $sPaymentMethodName];
     }
 
     /**
      * Sets the saveCheckoutFields flag
      *
-     * @param int $iSave value 1 if checkout data should be saved 0 if not
+     * @param int    $iSave value 1 if checkout data should be saved 0 if not
+     * @param string $sPaymentMethodName
      *
      * @since 1.2.0
      */
-    public static function setSaveCheckoutFields($iSave)
+    public static function setSaveCheckoutFields($iSave, $sPaymentMethodName)
     {
         $oSession = Registry::getConfig()->getSession();
         $aDynvalues = $oSession->getVariable('dynvalue');
-        $aDynvalues['saveCheckoutFields'] = $iSave;
+        $aDynvalues['saveCheckoutFields' . $sPaymentMethodName] = $iSave;
 
         $oSession->setVariable('dynvalue', $aDynvalues);
     }
